requirements.txt
numpy~=1.18.1
python-docx~=0.8.10
beautifulsoup4~=4.8.2
seaborn~=0.10.0
matplotlib~=3.1.3
pandas~=1.0.1
pillow~=7.0.0